fix: 修复 UserThirdpartyBinding 外键约束类型不匹配问题
- User.ID 和 UserThirdpartyBinding.UserID 统一使用 snowflake.ID(bigint unsigned)
- 移除 handlers 中不必要的 uint64() 类型转换
- 保持项目所有雪花 ID 字段类型一致性

+++ src/contexts/ordercontext/application/handlers/miniprogram_auth.go
@@ -195,7 +195,7 @@ func (h *MiniProgramAuthHandler) findOrCreateUser(sessionInfo *wechat.SessionInf
 
 	// 创建绑定
 	binding = &models.UserThirdpartyBinding{
-		UserID:         uint64(user.ID),
+		UserID:         user.ID,
 		Provider:       oauth.ProviderWeChat.String(),
 		ProviderUserID: providerUserID,
 		UnionID:        sessionInfo.UnionID,
